web-app/backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.0.0
channels_redis==4.2.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.0.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
zope.interface==7.2
psycopg2-binary>=2.9
